Fix and simplify imports
