+++ tests/tree_hash/test_containers.py
@@ -1,5 +1,8 @@
 import pytest
 
+from ssz.hash import (
+    hash_eth2,
+)
 from ssz.sedes import (
     List,
     Serializable,
@@ -7,10 +10,7 @@ from ssz.sedes import (
     uint32,
     uint32_list,
 )
-from ssz.tree_hash_backup.hash_eth2 import (
-    hash_eth2,
-)
-from ssz.tree_hash_backup.tree_hash import (
+from ssz.tree_hash import (
     hash_tree_root,
 )
